Add a BQ transfer job for releasetest
