+++ cron/releasetest/k8s/transfer.yaml
@@ -1,0 +1,41 @@
+# Copyright 2021 Security Scorecard Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: batch/v1beta1
+kind: CronJob
+metadata:
+  name: scorecard-bq-transfer-releasetest
+spec:
+  # Daily at 9am and 6pm PST.
+  schedule: "0 1,16 * * *"
+  concurrencyPolicy: "Forbid"
+  jobTemplate:
+    spec:
+      template:
+        spec:
+          restartPolicy: OnFailure
+          containers:
+          - name: bq-transfer-releasetest
+            image: gcr.io/openssf/scorecard-bq-transfer:latest
+            imagePullPolicy: Always
+            env:
+            - name: SCORECARD_DATA_BUCKET_URL
+              value: "gs://ossf-scorecard-data-releasetest"
+            - name: SCORECARD_BIGQUERY_TABLE
+              value: "scorecard_releasetest"
+            resources:
+              limits:
+                memory: 1Gi
+              requests:
+                memory: 1Gi
